Modify: 검색 결과 분리
