+++ package.json
@@ -22,7 +22,7 @@
 		"react-icons": "^4.10.1",
 		"react-lazy-load-image-component": "^1.6.0",
 		"react-redux": "^8.1.0",
-		"react-router-dom": "^6.12.1",
+		"react-router-dom": "^6.25.0",
 		"react-scripts": "^5.0.1",
 		"react-slick": "^0.29.0",
 		"redux": "^4.2.1",
@@ -53,6 +53,7 @@
 		]
 	},
 	"devDependencies": {
+		"@types/react-router-dom": "^5.3.3",
 		"@typescript-eslint/eslint-plugin": "^5.59.11",
 		"@typescript-eslint/parser": "^5.59.11",
 		"autoprefixer": "^10.4.19",
