+++ src/components/searchBar.tsx
@@ -1,49 +1,29 @@
 import React, { useState, FormEvent } from 'react';
+import { useNavigate } from 'react-router-dom';
 import SearchIcon from '@mui/icons-material/Search';
 import axios from 'axios';
 
-interface Level {
-	level: number;
-	color: string;
-}
-
-interface SearchResult {
-	name: string;
-	addressRoad: string;
-	addressLot: string;
-	latitude: string;
-	longitude: string;
-	levelList: Level[];
-	logoUrl: string;
-}
-
 const SearchBar = () => {
-	const [query, setQuery] = useState<string>(''); // 검색어 상태를 관리
-	const [data, setData] = useState<SearchResult[]>([]); // 데이터를 관리
-	const [error, setError] = useState<string | null>(null); // 에러 상태를 관리
-	const [loading, setLoading] = useState<boolean>(false); // 로딩 상태를 관리
+	// 검색어 상태 관리
+	const [query, setQuery] = useState<string>('');
+	const navigate = useNavigate();
 
 	const handleSearch = async (event: FormEvent) => {
-		event.preventDefault(); // 페이지 새로 고침을 방지
-
-		setLoading(true); // 검색 시작 시 로딩 상태로 변경
-		setError(null); // 기존 에러 초기화
+		event.preventDefault(); // 기본 폼 동작 방지
 
 		try {
 			const res = await axios.get('http://localhost:8080/climbing-info', {
 				params: {
 					searchKey: 'name',
-					searchValue: query, // 검색어(query)를 서버로 전송
+					searchValue: query,
 				},
 			});
 
 			console.log('서버 응답 데이터:', res.data);
-			setData(res.data);
+			// 검색 결과를 detail 페이지로 네비게이션
+			navigate('/detail', { state: { searchResults: res.data } });
 		} catch (err) {
 			console.error('데이터를 가져오는 중 에러 발생:', err);
-			setError('데이터를 가져오는 중 에러가 발생했습니다.');
-		} finally {
-			setLoading(false); // 검색 완료 시 로딩 상태 해제
 		}
 	};
 
@@ -67,43 +47,6 @@ const SearchBar = () => {
 					검색
 				</button>
 			</form>
-
-			<div className="mt-5">
-				{loading && <p>로딩 중...</p>}
-				{error && <p className="text-red-500">{error}</p>}
-				{Array.isArray(data) && data.length > 0 ? (
-					<ul>
-						{data.map((item) => (
-							<li key={item.name}>
-								<div className="flex items-center">
-									<img
-										src={item.logoUrl}
-										alt={item.name}
-										className="w-16 h-16 mr-4"
-									/>
-									<div>
-										<p className="font-bold">{item.name}</p>
-										<p>{item.addressRoad}</p>
-										<p>{item.addressLot}</p>
-										<div className="flex">
-											{Array.isArray(item.levelList) &&
-												item.levelList.map((level) => (
-													<div
-														key={level.level}
-														style={{ backgroundColor: level.color }}
-														className="w-4 h-4 rounded-full mr-1"
-													/>
-												))}
-										</div>
-									</div>
-								</div>
-							</li>
-						))}
-					</ul>
-				) : (
-					<p>검색 결과가 없습니다.</p>
-				)}
-			</div>
 		</div>
 	);
 };
