+++ src/pages/main.tsx
@@ -1,12 +1,10 @@
 import React from 'react';
 import RandomImg from '../components/randomImg';
-import SearchBar from '../components/searchBar';
 
 const Main = () => {
 	return (
 		<div>
 			<RandomImg />
-			<SearchBar />
 		</div>
 	);
 };
